requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
pytest-django==4.8.0
pytest-mock==3.14.0
psycopg2==2.9.9
django-jazzmin==3.0.0
drf-yasg==1.21